ci: refactor into coverage
